+++ shared/src/main/scala/amf/client/remod/amfcore/plugins/validate/ValidationConfiguration.scala
@@ -15,6 +15,6 @@ case class ValidationConfiguration(amfConfig: AMFGraphConfiguration) {
 
   val eh: AMFErrorHandler                              = amfConfig.errorHandlerProvider.errorHandler()
   val executionContext: ExecutionContext               = amfConfig.getExecutionContext
-  val maxYamlReferences: Option[Long]                  = amfConfig.options.parsingOptions.maxYamlReferences
+  val maxYamlReferences: Option[Int]                   = amfConfig.options.parsingOptions.maxYamlReferences
   val constraints: Map[ProfileName, ValidationProfile] = amfConfig.registry.constraintsRules
 }
